+++ packages/togetherai/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @ai-sdk/togetherai
 
+## 3.0.13
+
+### Patch Changes
+
+- Updated dependencies [cd06458]
+  - @ai-sdk/provider-utils@5.0.11
+  - @ai-sdk/openai-compatible@3.0.12
+
+## 3.0.12
+
+### Patch Changes
+
+- Updated dependencies [0b61267]
+  - @ai-sdk/openai-compatible@3.0.11
+
 ## 3.0.11
 
 ### Patch Changes